+++ packages/entities/entities-routes/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.10.17](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-routes@3.10.16...@kong-ui-public/entities-routes@3.10.17) (2025-03-20)
+
+**Note:** Version bump only for package @kong-ui-public/entities-routes
+
+
+
+
+
+## [3.10.16](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-routes@3.10.15...@kong-ui-public/entities-routes@3.10.16) (2025-03-20)
+
+**Note:** Version bump only for package @kong-ui-public/entities-routes
+
+
+
+
+
+## [3.10.15](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-routes@3.10.14...@kong-ui-public/entities-routes@3.10.15) (2025-03-20)
+
+**Note:** Version bump only for package @kong-ui-public/entities-routes
+
+
+
+
+
+## [3.10.14](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-routes@3.10.13...@kong-ui-public/entities-routes@3.10.14) (2025-03-18)
+
+**Note:** Version bump only for package @kong-ui-public/entities-routes
+
+
+
+
+
 ## [3.10.13](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-routes@3.10.12...@kong-ui-public/entities-routes@3.10.13) (2025-03-17)